Move typecheck def to decorators
