+++ python/hsfs/decorators.py
@@ -15,6 +15,7 @@
 #
 
 import functools
+import os
 
 
 def not_connected(fn):
@@ -53,3 +54,16 @@ class NoHopsworksConnectionError(Exception):
         super().__init__(
             "Connection is not active. Needs to be connected for feature store operations."
         )
+
+
+if os.environ.get("HOPSWORKS_RUN_WITH_TYPECHECK", False):
+    from typeguard import typechecked
+else:
+    from typing import TypeVar
+
+    _T = TypeVar("_T")
+
+    def typechecked(
+        target: _T,
+    ) -> _T:
+        return target if target else typechecked
