+++ python/hsfs/feature_store.py
@@ -46,10 +46,10 @@ from hsfs.core import (
     training_dataset_api,
     transformation_function_engine,
 )
+from hsfs.decorators import typechecked
 from hsfs.embedding import EmbeddingIndex
 from hsfs.statistics_config import StatisticsConfig
 from hsfs.transformation_function import TransformationFunction
-from hsfs.util import typechecked
 
 
 @typechecked
